+++ NEWS.md
@@ -1,10 +1,12 @@
 ## Version 1.0.1
 
-This is a patch release, resolving some problems with `plot_discrim()`
+This is a patch release, resolving some problems with `plot_discrim()` and adding functionality
 
 * Fixed buglet in `plot_discrim()` where the variables in a formula needed to be reversed.
 * `plot_discrim()` now gets the `data` from the object, if not supplied.
 * Added `tile.alpha` arg to `plot_discim()`; document how to customize colors, shapes, ...
+* `plot_discrim()` gains an `ellipse` argument to draw data ellipses for the groups.
+* Flesh out package description in README.
 
 ## Version 1.0.0